requirements.txt
essentials
numpy
pyyaml 
torch 
matplotlib 
scikit-learn
absl-py
dragonfly-opt